requirements.txt
python-jobspy>=1.1.70
pyyaml>=6.0
flask>=3.0
